requirements.txt

farm-haystack==1.18.0
pandas<2.0.0
pydantic==1.10.13
gradio==3.50.2
